+++ src/index.ts
@@ -6,7 +6,7 @@ export * from "./list";
 export * from "./order";
 export * from "./outline";
 export * from "./position";
-export * from "./position_char_map";
-export * from "./position_map";
-export * from "./position_set";
 export * from "./text";
+export * from "./unordered_collections/position_char_map";
+export * from "./unordered_collections/position_map";
+export * from "./unordered_collections/position_set";
