+++ src/unordered_collections/position_char_map.ts
@@ -1,28 +1,26 @@
-import { SerializedSparseArray, SparseArray } from "sparse-array-rled";
-import { ListSavedState } from "./list";
-import { Position } from "./position";
+import { SerializedSparseString, SparseString } from "sparse-array-rled";
+import { Position } from "../position";
+import { TextSavedState } from "../text";
 
 /**
- * A map from Positions to values of type `T`, **without ordering info**.
+ * A map from Positions to characters, **without ordering info**.
  *
- * This class is a simplified version of `List<T>` that does not consider the list order.
+ * This class is a simplified version of Text that does not consider the list order.
  * As a result, it does not require you to manage metadata, and it is slightly more efficient
- * than `List<T>`.
+ * than Text.
  *
- * For example, you can use a PositionMap to accumulate changes to save in a batch later.
+ * For example, you can use a PositionCharMap to accumulate changes to save in a batch later.
  * There the list order is unnecessary and managing metadata could be inconvenient.
- *
- * @typeParam T The value type.
  */
-export class PositionMap<T> {
+export class PositionCharMap {
   /**
-   * The internal state of this PositionMap: A map from bunchID
-   * to the [SparseArray](https://github.com/mweidner037/sparse-array-rled#readme)
+   * The internal state of this PositionCharMap: A map from bunchID
+   * to the [SparseString](https://github.com/mweidner037/sparse-array-rled#readme)
    * of values for that bunch.
    *
    * You are free to manipulate this state directly.
    */
-  readonly state: Map<string, SparseArray<T>>;
+  readonly state: Map<string, SparseString>;
 
   constructor() {
     this.state = new Map();
@@ -33,27 +31,27 @@ export class PositionMap<T> {
   // ----------
 
   /**
-   * Sets the value at the given position.
+   * Sets the char at the given position.
    */
-  set(pos: Position, value: T): void;
+  set(pos: Position, char: string): void;
   /**
-   * Sets the values at a sequence of Positions within the same [bunch](https://github.com/mweidner037/list-positions#bunches).
+   * Sets the chars at a sequence of Positions within the same [bunch](https://github.com/mweidner037/list-positions#bunches).
    *
    * The Positions start at `startPos` and have the same `bunchID` but increasing `innerIndex`.
    *
    * @see {@link expandPositions}
    */
-  set(startPos: Position, ...sameBunchValues: T[]): void;
-  set(startPos: Position, ...values: T[]): void {
+  set(startPos: Position, chars: string): void;
+  set(startPos: Position, chars: string): void {
     let arr = this.state.get(startPos.bunchID);
     if (arr === undefined) {
-      arr = SparseArray.new();
+      arr = SparseString.new();
     }
-    arr.set(startPos.innerIndex, ...values);
+    arr.set(startPos.innerIndex, chars);
   }
 
   /**
-   * Deletes the given position, making it and its value no longer present in the map.
+   * Deletes the given position, making it and its char no longer present in the list.
    */
   delete(pos: Position): void;
   /**
@@ -79,7 +77,7 @@ export class PositionMap<T> {
   }
 
   /**
-   * Deletes every position in the map, making it empty.
+   * Deletes every char in the list, making it empty.
    */
   clear() {
     this.state.clear();
@@ -90,9 +88,9 @@ export class PositionMap<T> {
   // ----------
 
   /**
-   * Returns the value at the given position, or undefined if it is not currently present.
+   * Returns the char at the given position, or undefined if it is not currently present.
    */
-  get(pos: Position): T | undefined {
+  get(pos: Position): string | undefined {
     return this.state.get(pos.bunchID)?.get(pos.innerIndex);
   }
 
@@ -108,16 +106,16 @@ export class PositionMap<T> {
   // ----------
 
   /**
-   * Iterates over [pos, value] pairs in the map, **in no particular order**.
+   * Iterates over [pos, char] pairs in the map, **in no particular order**.
    */
-  [Symbol.iterator](): IterableIterator<[pos: Position, value: T]> {
+  [Symbol.iterator](): IterableIterator<[pos: Position, char: string]> {
     return this.entries();
   }
 
   /**
-   * Iterates over [pos, value] pairs in the map, **in no particular order**.
+   * Iterates over [pos, char] pairs in the map, **in no particular order**.
    */
-  *entries(): IterableIterator<[pos: Position, value: T]> {
+  *entries(): IterableIterator<[pos: Position, char: string]> {
     for (const [bunchID, arr] of this.state) {
       for (const [innerIndex, value] of arr.entries()) {
         yield [{ bunchID, innerIndex }, value];
@@ -130,14 +128,14 @@ export class PositionMap<T> {
   // ----------
 
   /**
-   * Returns a saved state for this map, which is identical to its saved state as a `List<T>`.
+   * Returns a saved state for this map, which is identical to its saved state as a Text.
    *
-   * The saved state describes our current (Position -> value) map in JSON-serializable form.
-   * You can load this state on another PositionMap (or List) by calling `load(savedState)`,
+   * The saved state describes our current (Position -> char) map in JSON-serializable form.
+   * You can load this state on another PositionCharMap (or Text) by calling `load(savedState)`,
    * possibly in a different session or on a different device.
    */
-  save(): ListSavedState<T> {
-    const savedState: { [bunchID: string]: SerializedSparseArray<T> } = {};
+  save(): TextSavedState {
+    const savedState: { [bunchID: string]: SerializedSparseString } = {};
     for (const [bunchID, arr] of this.state) {
       if (!arr.isEmpty()) {
         savedState[bunchID] = arr.serialize();
@@ -147,16 +145,16 @@ export class PositionMap<T> {
   }
 
   /**
-   * Loads a saved state returned by another PositionMap's (or List's) `save()` method.
+   * Loads a saved state returned by another PositionCharMap's (or Text's) `save()` method.
    *
-   * Loading sets our (Position -> value) map to match the saved PositionMap, *overwriting*
+   * Loading sets our (Position -> char) map to match the saved PositionCharMap, *overwriting*
    * our current state.
    */
-  load(savedState: ListSavedState<T>): void {
+  load(savedState: TextSavedState): void {
     this.clear();
 
     for (const [bunchID, savedArr] of Object.entries(savedState)) {
-      this.state.set(bunchID, SparseArray.deserialize(savedArr));
+      this.state.set(bunchID, SparseString.deserialize(savedArr));
     }
   }
 }
